+++ packages/orq-rc/RELEASES.md
@@ -9359,3 +9359,13 @@ Based on:
 - [python v4.12.0-rc.8] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.8 - packages/orq-rc
+
+## 2026-07-08 21:21:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.0 (2.916.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.9] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.9 - packages/orq-rc